Move read_templates to Config class
